changes, remove the specificProduct class and create new Supply class that represent a group of the same product arrived

+++ dev/Inventory Module/Business_Layer/Discount.java
@@ -5,9 +5,9 @@ import java.util.Date;
 public class Discount {
     private Date start_date;
     private Date end_date;
-    private int discount_percentage;
+    private double discount_percentage;
 
-    public Discount(Date start_date, Date end_date, int discount_percentage) {
+    public Discount(Date start_date, Date end_date, double discount_percentage) {
         this.start_date = start_date;
         this.end_date = end_date;
         this.discount_percentage = discount_percentage;
@@ -29,7 +29,7 @@ public class Discount {
         return end_date;
     }
 
-    public int getDiscount_percentage() {
+    public double getDiscount_percentage() {
         return discount_percentage;
     }
